% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{callDMGaussian}
\alias{callDMGaussian}
\title{callDMGaussian}
\usage{
callDMGaussian(inputMatrix, dm_results)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Calls DMs identified using Gaussian decomposition of the regions of interest. Useful but not as sensitive
}
\keyword{cats}
