% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{collapseChrom3}
\alias{collapseChrom3}
\title{collapseChrom3}
\usage{
collapseChrom3(
  inputMatrix,
  minimumSegments = 40,
  summaryFunction = cutAverage,
  logTrans = TRUE,
  binExpand = 1,
  minimumChromValue = 2
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Deprecated
}
\examples{
cat_function()
}
\keyword{cats}
