% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{collapseChromN}
\alias{collapseChromN}
\title{collapseChromN}
\usage{
collapseChromN(
  inputMatrix,
  summaryFunction = cutAverage,
  logTrans = FALSE,
  binExpand = 1,
  minimumChromValue = 2
)
}
\arguments{
\item{inputMatrix}{normalised input matrix}

\item{summaryFunction}{function to use to collapse chromoseome}

\item{logTrans}{is the data log-transformed - TRUE or FALSE (default: FALSE)}

\item{binExpand}{- are we collapsing adjacent bins? >1 = true; integer}

\item{minimumChromValue}{- minimum cutoff to keep a chromosome arm as informative}
}
\description{
This function collapses bins into chromosomes.
}
\keyword{CNV}
