% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{getLOHRegions2}
\alias{getLOHRegions2}
\title{getLOHRegions2}
\usage{
getLOHRegions2(
  inputMatrixIn,
  lossCutoff = (-0.25),
  uncertaintyCutLoss = 0.5,
  diffThreshold = 0.7,
  minLength = 3e+06,
  minSeg = 3,
  lossCutoffCells = 100,
  targetFun = IQR,
  signalBoost = 1e-06
)
}
\description{
LOH analysis - alternate approach
}
\keyword{LOH}
