% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{graphSmoothDM}
\alias{graphSmoothDM}
\title{graphSmoothDM}
\usage{
graphSmoothDM(clusterOutputs, origData, dm_outputs, numNN = 0, perplex = 20)
}
\description{
Louvain cluster smooth the DMs
}
\keyword{smoothing}
