% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{makeFakeCells}
\alias{makeFakeCells}
\title{Simulate pseudodiploid control cells}
\usage{
makeFakeCells(x, num = 300, sd_fact = 0.1)
}
\arguments{
\item{x}{Mean value to simulate around.}

\item{num}{Number of cells to simulate.}

\item{sd_fact}{Standard deviation as a fraction of \code{x}.}
}
\value{
A numeric vector of length \code{num}.
}
\description{
Draws normally distributed values around \code{x} to act as a synthetic diploid
baseline when a sample contains too few non-neoplastic cells.
}
\seealso{
Other CNV calling: 
\code{\link{clusterCNV}()},
\code{\link{getAlteredSegments}()},
\code{\link{identifyCNVClusters}()}
}
\concept{CNV calling}
