% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{normalizeMatrix}
\alias{normalizeMatrix}
\title{normalizeMatrix}
\usage{
normalizeMatrix(
  inputMatrix,
  logNorm = TRUE,
  maxZero = 500,
  imputeZeros = FALSE,
  blacklistProp = 0.8,
  priorCount = 1,
  blacklistCutoff = 0,
  dividingFactor = 1e+06
)
}
\description{
This function normalises the input matrix file. Deprecated. Use normalizeMatrixN instead
}
\keyword{cats}
