% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{scaleMatrix}
\alias{scaleMatrix}
\title{Scale arm signal to Z-scores}
\usage{
scaleMatrix(inputMatrix, median_iqr_list, spread = FALSE)
}
\arguments{
\item{inputMatrix}{Collapsed, filtered matrix.}

\item{median_iqr_list}{List of centres and IQRs from \code{\link[=computeCenters]{computeCenters()}}.}

\item{spread}{Return long format (\code{Cell}, \code{Density}) instead of wide.}
}
\value{
The scaled matrix, wide or long depending on \code{spread}.
}
\description{
Centres and scales each arm using the centres and IQRs from
\code{\link[=computeCenters]{computeCenters()}}, and drops \code{chrYp}.
}
\seealso{
Other normalisation: 
\code{\link{collapseChrom3N}()},
\code{\link{computeCenters}()},
\code{\link{normalizeMatrixN}()}
}
\concept{normalisation}
