% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{scaleMatrixBins}
\alias{scaleMatrixBins}
\title{Merge adjacent genomic bins}
\usage{
scaleMatrixBins(inputMatrix, binSizeRatio, inColumn)
}
\arguments{
\item{inputMatrix}{Matrix with \code{chrom} and a position column.}

\item{binSizeRatio}{Factor by which to widen bins. \code{2} doubles bin width.}

\item{inColumn}{Name of the position column to rebin.}
}
\value{
The matrix regrouped on the widened bins, with numeric columns
summed.
}
\description{
Merge adjacent genomic bins
}
\seealso{
Other summarisation: 
\code{\link{cutAverage}()}
}
\concept{summarisation}
