% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{annotateCNV3}
\alias{annotateCNV3}
\title{annotateCNV3}
\usage{
annotateCNV3(cnvResults, saveOutput = TRUE, maxClust2 = 4, outputSuffix = "_1")
}
\arguments{
\item{cnvResults}{output of filtered CNVs (set of lists)}

\item{saveOutput}{save output to a file on disk}

\item{maxClust2}{maximum # of clusters}

\item{outputSuffix}{output suffix to save files}
}
\description{
Annotates the filtered CNV calls and saves the results, and estimates absolute copy numbers
}
\examples{
annotateCNV3(cleanCNV,saveOutput=TRUE,outputSuffix="_nolog")
}
\keyword{CNV}
\keyword{output}
