% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{annotateCNV4B}
\alias{annotateCNV4B}
\title{annotateCNV4B}
\usage{
annotateCNV4B(
  cnvResults,
  expectedNormals,
  saveOutput = TRUE,
  maxClust2 = 4,
  outputSuffix = "_1",
  sdCNV = 0.6,
  filterResults = TRUE,
  filterRange = 0.8,
  minAlteredCellProp = 0.75
)
}
\arguments{
\item{cnvResults}{output of filtered CNVs (set of lists)}

\item{expectedNormals}{list of normal barcodes}

\item{saveOutput}{save output to a file on disk}

\item{maxClust2}{maximum # of clusters}

\item{outputSuffix}{output suffix to save files}

\item{sdCNV}{sd for CNV copy number estimation}

\item{filterResults}{filter results to remove likely unaltered chromosomes}

\item{filterRange}{copy number range minimum}

\item{minAlteredCellProp}{filtering parameter - proportion of normal cell population that should be unaltered, default is 0.75}
}
\description{
Annotates the filtered CNV calls and saves the results, and estimates absolute copy numbers
unlike the regular version, this uses predicted normal cells to set the normal values to 2
}
\keyword{CNV}
\keyword{output}
