% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{annotateLosses}
\alias{annotateLosses}
\title{AnnotateLosses}
\usage{
annotateLosses(lossFile)
}
\arguments{
\item{lossFile}{the file for LOH putative reginos}
}
\description{
This function allows you to annotate loss regions
}
\keyword{Losses}
