% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{clusterCNV}
\alias{clusterCNV}
\title{clusterCNVs}
\usage{
clusterCNV(initialResultList, medianIQR, maxClust = 4, minDiff = 0.25)
}
\description{
Cleans the putative CNV calls
}
\keyword{CNV}
