% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{clusterCNV}
\alias{clusterCNV}
\title{clusterCNVs}
\usage{
clusterCNV(initialResultList, medianIQR, maxClust = 4, minDiff = 0.25)
}
\arguments{
\item{initialResultList}{output list generated by identifyCNVClusters}

\item{medianIQR}{median and IQR list of sample}

\item{maxClust}{maximum # of clusters fed into identifyCNVClusters}

\item{minDiff}{minimum Z-score difference between clusters}
}
\description{
Cleans the putative CNV calls
}
\keyword{CNV}
