% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{collapseChrom3N}
\alias{collapseChrom3N}
\title{collapseChrom3N}
\usage{
collapseChrom3N(
  inputMatrix,
  minimumSegments = 40,
  summaryFunction = cutAverage,
  logTrans = FALSE,
  binExpand = 1,
  minimumChromValue = 2,
  tssEnrich = 5,
  logBase = 2,
  minCPG = 300
)
}
\arguments{
\item{tssEnrich}{multiply CpG by weighting factor (default 1)}

\item{logBase}{log base to apply if doing a log weighting of the CpG scaling factor}

\item{minCPG}{keep only regions with at least CpG score of X (300)}

\item{inputMatrix:}{normalised matrix}

\item{summaryFunction:}{function to use to summarise signal}

\item{logTrans:}{whether matrix is log-transformed}

\item{binExpand:}{whether to merge adjacent bins, if so how many (merges if > 1)}

\item{minimumChromValue:}{cutoff to keep a chromosome}
}
\description{
This function summarises signal across chromosomes and normalises using the CpG content
}
\keyword{CNV}
