% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{computeCenters}
\alias{computeCenters}
\title{computeCenters}
\usage{
computeCenters(inputMatrix, summaryFunction = cutAverage)
}
\arguments{
\item{inputMatrix}{- normalised, filtered matrix}

\item{summaryFunction}{- summary function to use (defaults to cutAverage)}
}
\description{
Recompute range values after normalization and filtering. Returns a list containing values with summary values and IQRs
}
\examples{
listCenters <- computeCenters(inputMatrix,cutAverage)
}
\keyword{CNV}
\keyword{median}
