% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{cutAverage}
\alias{cutAverage}
\title{cutAverage}
\usage{
cutAverage(inputVector)
}
\arguments{
\item{inputVector}{- vector of numbers}
}
\description{
Averages regions of vector between prespecified quantiles
}
\keyword{average`}
