% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{estimateCellCycleFraction}
\alias{estimateCellCycleFraction}
\title{estimateCellCycleFraction
Detect cycling cells using chromosome X
Returns list of barcodes and cluster assignments, and list of non-neoplastic cells}
\usage{
estimateCellCycleFraction(
  inputMatrix,
  sampName,
  cutoff = 10000,
  maxG = 4,
  logTrans = FALSE,
  modelName = "E",
  chromToUse = "chrX"
)
}
\arguments{
\item{inputMatrix}{Matrix (unfiltered before normalization only)}

\item{cutoff}{Minimum cutoff to discard noisy cell}

\item{maxG}{maximum number of G components to use (default is 4)}

\item{logTrans}{log transform data (default is faulse)}

\item{modelName}{Model option for MCLUST (default is E)}

\item{chromToUse}{Chromosome to use for normalization (default is chrX)}
}
\description{
estimateCellCycleFraction
Detect cycling cells using chromosome X
Returns list of barcodes and cluster assignments, and list of non-neoplastic cells
}
\keyword{CNV}
\keyword{cell}
\keyword{cycle}
