% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{getAlteredSegments}
\alias{getAlteredSegments}
\title{getAlteredSegments}
\usage{
getAlteredSegments(
  inputMatrix,
  clusterResults,
  minSeglen = 3,
  lossThreshold = 0.5,
  gainThreshold = 1.6,
  minCutoff = 50,
  rangeThreshold = 0.4
)
}
\arguments{
\item{inputMatrix}{Unprocessed and normalized input matrix}

\item{clusterResults}{Result list of identifyNonNeoplastic}

\item{minSeglen}{Minimum segment length for changepoint analysis}

\item{lossThreshold}{fold change threshold to call something a relative loss (default 0.5)}

\item{gainThreshold}{fold change threshold to  call something a relative gain (default 1.6)}

\item{minCutoff}{Number of altered cells in smallest group to keep a putative alteration}

\item{rangeThreshold}{threshold filter between max and min fold change to keep an alteration (Default 0.4)}
}
\description{
This function relies on pre-determined 'normal' cell set to compute smaller regions of altered accessibility
}
\keyword{CNV}
\keyword{comparison}
