% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{getDoubleMinutes}
\alias{getDoubleMinutes}
\title{getDoubleMinutes}
\usage{
getDoubleMinutes(
  inputMatrix,
  targetCell,
  doPlot = FALSE,
  penalty_type = "SIC",
  doLosses = FALSE,
  peakCutoff = 3,
  lossCutoff = -1
)
}
\description{
This function allows you to express your love of cats.
}
\keyword{double}
\keyword{minutes}
