% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{getLOHRegions}
\alias{getLOHRegions}
\title{getLOHRegions}
\usage{
getLOHRegions(
  inputMatrixIn,
  lossCutoff = (-0.25),
  uncertaintyCutLoss = 0.5,
  diffThreshold = 0.9,
  minLength = 3e+06,
  minSeg = 3,
  lossCutoffCells = 100,
  targetFun = IQR,
  quantileLimit = 0.3,
  cpgCutoff = 0,
  meanThreshold = 4,
  dummyQuantile = 0.5,
  dummyPercentile = 0.2,
  dummySd = 0.2
)
}
\arguments{
\item{lossCutoff}{- minimum Z score for losses (default is -0.25)}

\item{uncertaintyCutLoss}{- maximum uncertainty to accept a cell assignment}

\item{diffThreshold}{- difference threshold between bins}

\item{minLength}{- minimum length of a LOH region}

\item{minSeg}{- minimum number of bins for changepoint analysis}

\item{lossCutoffCells}{- minimum number of cells to accept a LOH (default is 100)}

\item{targetFun}{- target function to identify LOH regions}

\item{quantileLimit}{quantile of signal per cell population to use to identify losses in changepoints (default is 0.3)}

\item{inputMatrix}{- normalised input fragment count matrix}

\item{signalBoost}{- value to add prior to logtransforming signal}

\item{lossCutoffReads}{minimum number of cells showing alteration to keep in final list (filter)}
}
\description{
Calls putative regions of LOH
}
\keyword{CNV}
\keyword{LOH}
