% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{graphCNVDistribution}
\alias{graphCNVDistribution}
\title{graphCNVDistribution}
\usage{
graphCNVDistribution(inputMatrix, outputSuffix = "_all")
}
\arguments{
\item{inputMatrix}{(the normalised matrix)}

\item{outputSuffix}{suffix to append to pdf output file}
}
\description{
Makes graph of normalised, collapsed fragment data
}
\keyword{output}
