% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{identifyCNVClusters}
\alias{identifyCNVClusters}
\title{identifyCNVClusters}
\usage{
identifyCNVClusters(
  inputMatrix,
  median_iqr,
  useDummyCells = FALSE,
  propDummy = 0.1,
  maxClust = 4,
  deltaMean = 0.1,
  minDiff = 0.25,
  deltaBIC2 = 50,
  bicMinimum = 5,
  minMix = 0.3,
  subsetSize = 500,
  fakeCellSD = 0.2,
  uncertaintyCutoff = 1,
  summaryFunction = cutAverage,
  mergeCutoff = 0.5
)
}
\arguments{
\item{x}{(mean of values)}

\item{num}{(number of cells)}

\item{sd_fact}{(scaling factor for cell SD)}
}
\description{
How the sausage is made - uses Gaussian decomposition to analyse a filtered, normalised input matrix and call putative CNVs
}
