% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{identifyCNVClusters}
\alias{identifyCNVClusters}
\title{identifyCNVClusters}
\usage{
identifyCNVClusters(
  inputMatrix,
  median_iqr,
  useDummyCells = TRUE,
  propDummy = 0.25,
  deltaMean = 0.03,
  subsetSize = 500,
  fakeCellSD = 0.1,
  summaryFunction = cutAverage,
  minDiff = 0.25,
  deltaBIC2 = 0.25,
  bicMinimum = 0.1,
  minMix = 0.3,
  uncertaintyCutoff = 0.55,
  mergeCutoff = 3,
  summarySuffix = "",
  shrinky = 0,
  maxClust = 4,
  IQRCutoff = 0.25,
  medianQuantileCutoff = 0.3,
  normalCells = NULL
)
}
\arguments{
\item{median_iqr}{matrix of medians and IQRs}

\item{useDummyCells}{use pseudodiploid control}

\item{propDummy}{pseudodiploid control as proportion of normal cells (0.25 = 25 percent)}

\item{deltaMean}{minimum difference between clusters}

\item{subsetSize}{subset of cells to use for initial computation (useful in large samples)}

\item{fakeCellSD}{SD of dummy cells as proportion of range of normal cells (0.10 is default, 0.10-0.20 works best in most cases)}

\item{summaryFunction}{summarising function for chromosome arms (default is trimmed average)}
}
\description{
How the sausage is made - uses Gaussian decomposition to analyse a filtered, normalised input matrix and call putative CNVs
}
