% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{identifyDoubleMinutes}
\alias{identifyDoubleMinutes}
\title{identifyDoubleMinutes}
\usage{
identifyDoubleMinutes(
  inputMatrix,
  minCells = 100,
  qualityCutoff2 = 100,
  peakCutoff = 5,
  lossCutoff = -1,
  doPlots = FALSE,
  imageNumber = 1000,
  logTrans = FALSE,
  cpgTransform = FALSE,
  doLosses = FALSE,
  minThreshold = 4
)
}
\arguments{
\item{inputMatrix}{normalized input matrix}

\item{minCells}{minimum cells to retain a DM}

\item{qualityCutoff2}{minimum cells to retain DM (filter 2)}

\item{peakCutoff}{Z score minimum to retain a putative DM/amplification}

\item{doPlots}{Plot occasional cells as QC}

\item{imageNumber}{Plot every Nth cell (use in conjunction with doPlots = TRUE)}

\item{minThreshold}{Z-score maximum to continue processing cells (can save time in larger datasets; default 4)}
}
\description{
Identify double minutes on all chromosomes using changepoint analysis
}
\keyword{amplification}
\keyword{double}
\keyword{ecDNA}
\keyword{minutes}
