% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{identifyDoubleMinutes}
\alias{identifyDoubleMinutes}
\title{identifyDoubleMinutes}
\usage{
identifyDoubleMinutes(
  inputMatrix,
  minCells = 20,
  qualityCutoff2 = 40,
  logTrans = FALSE,
  cpgTransform = FALSE,
  peakCutoff = 3,
  lossCutoff = -1,
  doLosses = FALSE,
  doPlots = TRUE,
  imageNumber = 250
)
}
\description{
Identify double minutes on all chromosomes using changepoint analysis
}
\keyword{amplification}
\keyword{double}
\keyword{ecDNA}
\keyword{minutes}
