% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{identifyNonNeoplastic}
\alias{identifyNonNeoplastic}
\title{identifyNonNeoplastic
Detect neoplastic vs neoplastic cells using unsupervised clustering
Returns list of barcodes and cluster assignments, and list of non-neoplastic cells}
\usage{
identifyNonNeoplastic(
  inputMatrix,
  estimatedCellularity = 0.8,
  nmfComponents = 5,
  outputHeatmap = TRUE,
  cutHeight = 0.6,
  methodHclust = "ward.D"
)
}
\arguments{
\item{inputMatrix}{Matrix (after filterCells is called)}

\item{estimatedCellularity}{Expected cellularity of tumour (default 0.8)}

\item{nmfComponents}{# of components to use for NMF/SVD}

\item{outputHeatmap}{Output a heatmap of the clustering (default is yes)}

\item{cutHeight}{Cutting height as percent of max for dendrogram}

\item{methodHclust}{Method for clustering (default is Ward.D)}
}
\description{
identifyNonNeoplastic
Detect neoplastic vs neoplastic cells using unsupervised clustering
Returns list of barcodes and cluster assignments, and list of non-neoplastic cells
}
\keyword{CNV}
\keyword{clustering}
