% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{initialiseEnvironment}
\alias{initialiseEnvironment}
\title{initialiseEnvironment}
\usage{
initialiseEnvironment(
  genomeFile,
  cytobandFile,
  cpgFile,
  binSize = 1e+06,
  minFrags = 10000,
  cellSuffix = c("-1"),
  lowerTrim = 0.5,
  upperTrim = 0.8
)
}
\arguments{
\item{genomeFile}{Chromosome size file for genome of interest}

\item{cytobandFile}{Processed cytoband file}

\item{cpgFile}{Processed CpG island file}

\item{binSize}{Size of chromosome bins (defaults to 1e6)}

\item{minFrags}{Minimum # of fragments per cell (default 1e4)}

\item{cellSuffix}{Suffix appended to cells}

\item{lowerTrim}{Lower quantile trim for trimmed mean (default 0.5)}

\item{upperTrim}{Upper quantile trim for trimmed mean (default 0.8)}
}
\description{
This function prepares and preloads the genome data required for CNV analysis.
}
\keyword{configuration}
