% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{makeFakeCells}
\alias{makeFakeCells}
\title{makeFakeCells}
\usage{
makeFakeCells(x, num = 300, sd_fact = 0.1)
}
\arguments{
\item{x}{(mean of values)}

\item{num}{(number of cells)}

\item{sd_fact}{(scaling factor for cell SD)}
}
\description{
simulate a pseudodiploid control
}
