% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{normalizeMatrixN}
\alias{normalizeMatrixN}
\title{normalizeMatrixN}
\usage{
normalizeMatrixN(
  inputMatrix,
  logNorm = FALSE,
  maxZero = 500,
  imputeZeros = FALSE,
  blacklistProp = 0.8,
  priorCount = 1,
  blacklistCutoff = 0,
  dividingFactor = 1e+06
)
}
\arguments{
\item{inputMatrix}{- processed fragment table from your input}

\item{logNorm}{- whether to log normalize input file. default: FALSE}

\item{maxZero}{- maximum # of zeros per cell after removal of uninformative regions}

\item{imputeZeros}{- default to FALSE; will impute cells with medians - don't use it}

\item{blacklistProp}{- proportion of cells that have zeros in an interval to consider that interval blacklisted/uninformative - lower = more stringency}

\item{priorCount}{- use if log normalizing data - priorCount to add to fields before normalisation}

\item{blacklistCutoff}{- minimum signal in a bin to count it as "empty" for the blacklist cutoff}

\item{dividingFactor}{- deprecated}
}
\description{
This function normalizes the input matrix. Do this first.
}
