% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{readInputTable}
\alias{readInputTable}
\title{readInputTable}
\usage{
readInputTable(inputFile, sep = "\\t")

readInputTable(inputFile, sep = "\\t")
}
\arguments{
\item{inputFile}{input count fragment matrix}

\item{sep}{separator (defaults to tab)}
}
\description{
This function prepares and preloads the genome data required for CNV analysis.
}
\keyword{CNV}
\keyword{files}
\keyword{import}
\keyword{initialisation}
\keyword{loading}
