% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{saveSummaryStats}
\alias{saveSummaryStats}
\title{saveSummaryStats}
\usage{
saveSummaryStats(outputSuffix = "_stats")
}
\arguments{
\item{outputSuffix}{suffix to append to output file names. defaults to _stats}
}
\description{
saves summary statistics about loaded files
}
\keyword{summary}
