% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{scaleMatrix}
\alias{scaleMatrix}
\title{scaleMatrix}
\usage{
scaleMatrix(inputMatrix, median_iqr_list, spread = FALSE)
}
\arguments{
\item{inputMatrix}{normalised, filtered matrix}

\item{median_iqr_list}{list containing summary lists with median/trimmed mean and IQR (computed by computeCenters)}

\item{spread}{return the list in Tidyverse "spread" format - for plotting}
}
\description{
Scales the normalised, filtered input matrix based on provided median and IQR data
}
\examples{
scaleMatrix(inputMatrix,medianIQR)
}
\keyword{CNV}
\keyword{scale}
