% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{scaleMatrixBins}
\alias{scaleMatrixBins}
\title{Scale matrix bins}
\usage{
scaleMatrixBins(inputMatrix, binSizeRatio, inColumn)
}
\arguments{
\item{inputMatrix}{matrix of input stuff}

\item{binSizeRatio}{ratio to expand by (2 = double bin size)}
}
\description{
Merges together adjacent windows
}
\examples{
cat_function()
}
\keyword{cats}
