% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scDataSamp}
\alias{scDataSamp}
\title{Sample processed matrix for testing}
\format{
data.frame with 3103 columns and 1371 rows
}
\source{
1371 adult GBM cells
}
\usage{
scDataSamp
}
\description{
Sample processed matrix for testing
}
\keyword{datasets}
