% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scDataSamp}
\alias{scDataSamp}
\title{scData sample matrix}
\format{
data.frame with 3103 columns and 1371 rows
}
\usage{
scDataSamp
}
\description{
1371 adult GBM cells
}
\details{
Sample processed matrix for testing

NBD
}
\keyword{datasets}
