% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scDataSampClusters}
\alias{scDataSampClusters}
\title{Sample processed cluster data frame for testing}
\format{
data.frame
}
\source{
adult GBM cells
}
\usage{
scDataSampClusters
}
\description{
Sample processed cluster data frame for testing
}
\keyword{datasets}
