% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{setOutputFile}
\alias{setOutputFile}
\title{setOutputFile}
\usage{
setOutputFile(outputDir, outputFileSuffix)
}
\arguments{
\item{outputDir}{Output directory}

\item{outputFileSuffix}{Suffix to add to output files}
}
\description{
This function prepares and preloads the genome data required for CNV analysis.
}
\keyword{configuration}
