% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{smoothClusters}
\alias{smoothClusters}
\title{smoothClusters
Read clusters from file and CNV calls from input or an external file
Note: assumes files are comma delimited
Returns a smoothed matrix}
\usage{
smoothClusters(
  inputClusters,
  inputCNVList,
  inputCNVClusterFile = "",
  percentPositive = 0.5,
  removeEmpty = TRUE
)
}
\arguments{
\item{inputClusters}{input cluster matrix}

\item{inputCNVList}{list of CNVs with copy number estimates (2 as normal)}

\item{inputCNVClusterFile}{alternate option to specify external comma delimited file for CNV calls}

\item{percentPositive}{minimum percentage of a cluster to use to call positive for CNV, default is 0.5 (50 percent)}
}
\description{
smoothClusters
Read clusters from file and CNV calls from input or an external file
Note: assumes files are comma delimited
Returns a smoothed matrix
}
\keyword{CNV}
\keyword{clusters}
