% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_caller_functions_packaged.R
\name{testOverlap}
\alias{testOverlap}
\title{Test overlap}
\usage{
testOverlap(interval.1, interval.2)
}
\arguments{
\item{interval.1}{(text interval)}

\item{interval.2}{(text interval)}
}
\description{
Tests overlap between two intervals
Internal use
}
